Model Refactor
Correct domain model package associations
