+++ src/main/java/com/howellsmith/oss/nfcpipass/dao/UserRepository.java
@@ -1,6 +1,6 @@
 package com.howellsmith.oss.nfcpipass.dao;
 
-import com.howellsmith.oss.nfcpipass.model.User;
+import com.howellsmith.oss.nfcpipass.model.domain.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
